+++ src/commandGroups/streams.ts
@@ -2,6 +2,7 @@ import { commandTable as echo} from "../commands/echo.js";
 import { commandTable as write } from "../commands/write.js";
 import { commandTable as json } from "../commands/json.js";
 import { commandTable as b64 } from "../commands/b64.js";
+import { commandTable as color } from "../commands/color.js";
 import { commandCollection } from "../tools/commandCollection.js";
 import { cmdTableToCommandCompounts } from "../tools/commandCreatorTools.js";
 
@@ -9,7 +10,8 @@ const commandTable = {
     ...echo,
     ...write,
     ...json,
-    ...b64
+    ...b64,
+    ...color
 };
 
 const compounds = cmdTableToCommandCompounts(commandTable);